ENH: move warnings context manager, rename
