+++ numpy/testing/tests/test_utils.py
@@ -1,10 +1,17 @@
 from __future__ import division, absolute_import, print_function
 
 import warnings
+from warnings import warn, simplefilter
 import sys
 
 import numpy as np
-from numpy.testing import *
+from numpy.testing import (assert_equal, assert_array_equal,
+                           assert_almost_equal,
+                           assert_array_almost_equal,
+                           build_err_msg, raises, assert_raises,
+                           assert_warns, assert_no_warnings, assert_allclose,
+                           assert_approx_equal, assert_array_almost_equal_nulp,
+                           assert_array_max_ulp, catch_and_clear_warnings, run_module_suite)
 import unittest
 
 class _GenericTest(object):
@@ -251,6 +258,7 @@ class TestArrayAlmostEqual(_GenericTest, unittest.TestCase):
         assert_array_almost_equal(a, b)
         assert_array_almost_equal(b, a)
         assert_array_almost_equal(b, b)
+
 
 class TestAlmostEqual(_GenericTest, unittest.TestCase):
     def setUp(self):
@@ -688,5 +696,67 @@ class TestULP(unittest.TestCase):
             self.assertRaises(AssertionError,
                                   lambda: assert_array_max_ulp(nan, nzero,
                                                                maxulp=maxulp))
+
+def assert_warn_len_equal(mod, n_in_context):
+    mod_warns = mod.__warningregistry__
+    # Python 3.4 appears to clear any pre-existing warnings of the same type,
+    # when raising warnings inside a catch_warnings block. So, there is a
+    # warning generated by the tests within the context manager, but no
+    # previous warnings.
+    if 'version' in mod_warns:
+        assert_equal(len(mod_warns), 2)  # including 'version'
+    else:
+        assert_equal(len(mod_warns), n_in_context)
+
+
+def _get_fresh_mod():
+    # Get this module, with warning registry empty
+    my_mod = sys.modules[__name__]
+    try:
+        my_mod.__warningregistry__.clear()
+    except AttributeError:
+        pass
+    return my_mod
+
+
+def test_catch_and_clear_warnings():
+    # Initial state of module, no warnings
+    my_mod = _get_fresh_mod()
+    assert_equal(getattr(my_mod, '__warningregistry__', {}), {})
+    with catch_and_clear_warnings(modules=[my_mod]):
+        simplefilter('ignore')
+        warn('Some warning')
+    assert_equal(my_mod.__warningregistry__, {})
+    # Without specified modules, don't clear warnings during context
+    with catch_and_clear_warnings():
+        simplefilter('ignore')
+        warn('Some warning')
+    assert_warn_len_equal(my_mod, 1)
+    # Confirm that specifying module keeps old warning, does not add new
+    with catch_and_clear_warnings(modules=[my_mod]):
+        simplefilter('ignore')
+        warn('Another warning')
+    assert_warn_len_equal(my_mod, 1)
+    # Another warning, no module spec does add to warnings dict, except on
+    # Python 3.4 (see comments in `assert_warn_len_equal`)
+    with catch_and_clear_warnings():
+        simplefilter('ignore')
+        warn('Another warning')
+    assert_warn_len_equal(my_mod, 2)
+
+
+class my_cacw(catch_and_clear_warnings):
+    class_modules = (sys.modules[__name__],)
+
+
+def test_catch_and_clear_warnings_inherit():
+    # Test can subclass and add default modules
+    my_mod = _get_fresh_mod()
+    with my_cacw():
+        simplefilter('ignore')
+        warn('Some warning')
+    assert_equal(my_mod.__warningregistry__, {})
+
+
 if __name__ == '__main__':
     run_module_suite()
